Fix logging bug

+++ src/utils.c
@@ -1,28 +1,30 @@
 #include "utils.h"
 
+#include <limits.h>
+
 #include <libopencm3/cm3/systick.h>
 #include <libopencm3/cm3/nvic.h>
 
-volatile uint32_t system_microseconds;
+volatile uint64_t system_microseconds;
 
 void sys_tick_handler(void) {
 	system_microseconds++;
 }
 
-void usleep(uint32_t delay) {
-	uint32_t wake = system_microseconds + delay;
+void usleep(uint64_t delay) {
+	uint64_t wake = system_microseconds + delay;
 	while (wake > system_microseconds);
 }
 
-void msleep(uint32_t delay) {
-	uint32_t i;
+void msleep(uint64_t delay) {
+	uint64_t i;
 	for (i = 0; i < delay; ++i) usleep(1000);
 }
 
-uint32_t get_time_ms(void) {
+uint64_t get_time_ms(void) {
 	return system_microseconds / 1000;
 }
 
-uint32_t get_time_since(uint32_t start_time) {
+uint64_t get_time_since(uint64_t start_time) {
 	return get_time_ms() - start_time;
 }
