+++ src/utils.h
@@ -3,9 +3,9 @@
 
 #include <stdint.h>
 
-extern void usleep(uint32_t delay);
-extern void msleep(uint32_t delay);
-extern uint32_t get_time_ms(void);
-extern uint32_t get_time_since(uint32_t start_time);
+extern void usleep(uint64_t delay);
+extern void msleep(uint64_t delay);
+extern uint64_t get_time_ms(void);
+extern uint64_t get_time_since(uint64_t start_time);
 
 #endif
